# backend/api/api.yaml
openapi: 3.0.3
info:
  title: gradewise-api-backend
  version: 1.0.0
servers:
  - url: /v1
paths:
  /greet:
    get:
      summary: Greet a user by name
      operationId: GreetUser
      parameters:
        - name: name
          in: query
          description: User's name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful greeting
          content:
            application/json:
              schema:
                type: string
                properties:
                  message:
                    type: string
                    example: Hello, Alice! Goodbye!
  /faculty/register:
    post:
      tags: [ Faculty ]
      summary: Register a new faculty account
      operationId: RegisterFaculty
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FacultyRegistrationRequest'
      responses:
        "201":
          description: Faculty created (verification required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacultyRegistrationResponse'
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /courses:
    post:
      tags: [ Courses ]
      summary: Create a course
      operationId: CreateCourse
      # security: [ bearerAuth: [] ]  # enable when JWT is ready
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreateRequest'
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseCreateResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          description: Faculty not verified
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        "409":
          description: Duplicate course offering
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        default:
          $ref: '#/components/responses/ErrorResponse'

    get:
      tags: [ Courses ]
      summary: List courses
      operationId: ListCourses
      # security: [ bearerAuth: [] ]
      parameters:
        - in: query
          name: mine
          description: Only courses I own or grade
          required: false
          schema: { type: boolean, default: true }
        - in: query
          name: term
          required: false
          schema: { type: string, example: "Fall 2025" }
        - in: query
          name: code
          required: false
          schema: { type: string, example: "CS101" }
      responses:
        "200":
          description: Courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
        "401":
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/ErrorResponse'

components:
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    ErrorResponse:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:
    FacultyRegistrationRequest:
      type: object
      required: [ email, password, fullName ]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        fullName: { type: string }
        institution: { type: string }

    FacultyRegistrationResponse:
      type: object
      required: [ facultyId, email, status ]
      properties:
        facultyId: { type: string, format: uuid }
        email: { type: string, format: email }
        status:
          type: string
          enum: [ PENDING_VERIFICATION, VERIFIED ]
    CourseCreateRequest:
      type: object
      required: [ code, term, title ]
      properties:
        code: { type: string, example: "CS101" }
        term: { type: string, example: "Fall 2025" }
        title: { type: string, example: "Intro to CS" }
        section: { type: string, nullable: true, example: "001" }
        startDate: { type: string, format: date, nullable: true }
        endDate: { type: string, format: date, nullable: true }

    CourseCreateResponse:
      type: object
      properties:
        courseId: { type: string, format: uuid }

    Course:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        term: { type: string }
        title: { type: string }
        section: { type: string, nullable: true }
        startDate: { type: string, format: date, nullable: true }
        endDate: { type: string, format: date, nullable: true }
        creatorFacultyId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }

    Error:
      type: object
      required: [ code, message ]
      properties:
        code: { type: string, example: BAD_REQUEST }
        message: { type: string, example: Invalid request }
        details:
          type: object
          additionalProperties: true